+++ CHANGELOG.md
@@ -10,4 +10,9 @@ First official release.
 Advanced Installer VSTS Task 1.2
 -----------------------
 
-Port task to use the  Powershell3 handler and VSTS task SDK.
+* Port task to use the  Powershell3 handler and VSTS task SDK.
+
+Advanced Installer VSTS Task 1.3
+-----------------------
+
+* Search Advanced Installer in PATH if not found on system.